refactor(flashcards): split import export tab panels
